chore: rename sandbox files
