+++ public/app/features/plugins/sandbox/pluginDependencies.ts
