+++ public/app/features/plugins/sandbox/sandboxPluginLoaderRegistry.ts
